requirements.txt
Cerberus>=1.3.5
duckdb>=0.10.0
pandas>=2.0.3
pydantic>=2.4.2
PyYAML>=6.0.1
PyYAML>=6.0.1
Requests>=2.31.0
rich>=13.6.0
typer>=0.9.0
urllib3>=2.0.6
zstandard>=0.22.0
pycountry
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
beautifulsoup4>=4.12.0